requirement.txt
mysql-connector-python
argparse
beautifulsoup4
tqdm
xmltodict